adopt extension installation in profiles in web
